# backend/openapi.yaml
openapi: "3.0.3"
info:
  title: OneFoodDialer Merchandise Service API
  description: |
    REST API for the Merchandise Procurement Portal.
    All endpoints require a **Keycloak Bearer** token unless stated otherwise.
    Routes are served under the `/api` prefix by Laravel.
  version: "1.0.0"
  contact:
    name: OneFoodDialer Engineering
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://kong.yourdomain.com
    description: Production (via Kong Gateway)
...
security:
  - bearerAuth: []
paths:
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List orders (buyer sees own, approver sees all)
      operationId: listOrders
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [pending, approved, in_progress, delivered, completed, issues]
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Orders]
      summary: Place a new order (buyer only)
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "422":
          description: Validation error
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/orders/{id}/issue:
    post:
      tags: [Orders]
      summary: Report issue on delivered order (buyer only)
      operationId: reportIssue
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
      responses:
        "200":
          description: Order refreshed
        "422":
          description: Order not in delivered status

  # ─── Approvals ─────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued JWT token
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string, example: "Unauthorized: missing bearer token" }
    Forbidden:
      description: Insufficient role permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string, example: "Forbidden: missing required role" }
              required: { type: array, items: { type: string } }
              user_roles: { type: array, items: { type: string } }
  schemas:
...
    ProductDetail:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        price: { type: number, format: float }
        image_url: { type: string, nullable: true }
        stock: { type: integer, nullable: true }
        category_id: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OrderItem:
      type: object
      properties:
        id: { type: integer }
        order_id: { type: integer }
        product_id: { type: integer }
        qty_requested: { type: integer }
        qty_approved: { type: integer, nullable: true }
        price: { type: number, format: float }
        product:
          $ref: "#/components/schemas/ProductDetail"
    Address:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: string }
        name: { type: string }
        phone: { type: string }
        address_line: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
...
    Order:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: string }
        status:
          type: string
          enum:
            - pending_approval
            - partially_approved
            - approved
            - rejected
            - processing
            - dispatched
            - in_transit
            - delivered
            - partially_received
            - completed
            - issue_reported
        total_amount: { type: number, format: float }
        address_id: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        address:
          $ref: "#/components/schemas/Address"
    CreateOrderRequest:
      type: object
      required: [items, address_id]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [product_id, qty]
            properties:
              product_id: { type: integer }
              qty: { type: integer, minimum: 1 }
        address_id: { type: integer }
...
    PaginatedResponse:
      type: object
      properties:
        current_page: { type: integer }
        last_page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
        next_page_url: { type: string, nullable: true }
        prev_page_url: { type: string, nullable: true }